# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "link-controller"
        ],
        "summary": "Get the home page.",
        "operationId": "home",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "link-controller"
        ],
        "summary": "Generate short URL for provided long URL.",
        "operationId": "shortenUrl",
        "parameters": [
          {
            "name": "submitForm",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SubmitForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/original/{shortUrl}": {
      "get": {
        "tags": [
          "link-controller"
        ],
        "summary": "Return the corresponding long URL.",
        "operationId": "returnOriginalUrl",
        "parameters": [
          {
            "name": "shortUrl",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SubmitForm": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      }
    }
  }
}